+++ tests/Exceptions/CouldNotSendNotificationTest.php
@@ -8,16 +8,12 @@ use AvtoDev\FirebaseNotificationsChannel\Tests\AbstractTestCase;
 use AvtoDev\FirebaseNotificationsChannel\Exceptions\CouldNotSendNotification;
 
 /**
- * Class CouldNotSendNotificationTest.
- *
  * @covers \AvtoDev\FirebaseNotificationsChannel\Exceptions\CouldNotSendNotification
  */
 class CouldNotSendNotificationTest extends AbstractTestCase
 {
     /**
      * Check exception message.
-     *
-     * @throws \SebastianBergmann\RecursionContext\InvalidArgumentException
      */
     public function testInvalidNotification(): void
     {